+++ apx_sdk/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added `verification_method_id` method to `DidKey` type.
+- Added `sign_object` function.
 
 ### Changed